requirements.txt
photoshop_python_api
PySimpleGUI